#266 Consolidate authentication APIs.
