+++ src/Microsoft.AspNet.Http.Core/HttpRequest.cs
@@ -1,9 +1,6 @@
 // Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
-using System.Collections;
-using System.Collections.Generic;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
@@ -12,8 +9,6 @@ namespace Microsoft.AspNet.Http
 {
     public abstract class HttpRequest
     {
-        // TODO - review IOwinRequest for properties
-
         public abstract HttpContext HttpContext { get; }
 
         /// <summary>
